Added "happy path end-to-end test"
This is the main happy path test that connects all the individual
v1.0 APIs together to ensure the return values and parameters are
properly constructed and designed to allow applications to use
the SDK to implement complete flows.

The change also includes corresponding README.md updates for
instructions to set up the target environment and drive the tests.

Change-Id: Id523f479434de3588ba9ba345c500c5d0519e7f9
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>