[WIP] Implement enroll function to work with COP
Note:  Work in progress but wanted to submit for review

This change set introduces a new member services client
implementation for communcating with the new Fabric COP
server.  Currently the only API implemented is enroll

Made this impl the default from utils.getMemberService()

Save and load certificates in PEM format. Needs a change
in endorser.go to take PEM instead of DER.

Change-Id: Ib5d0cb32391f30dc20fb8779cbcc615400c6a7b6
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>