WIP Move MemberServices out of HFC
To support the new member service aka COP,
MemberService.js has been moved to it's own
directory, hfc-cop, so that it can be published
as a separate npmjs package.  This also
required the remaining HFC code to be moved
to a new directory, hfc.  The build tasks
were modified to run against both directories.
The readthedocs files remain in the root
fabric-node-sdk directory.  FAB-1084

Change-Id: I16c85d2e3b73dca883b56e106e35bf4fceb97a8b
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>