NodeSDK updates for new protobufs
Updates include signing of the transaction
and allowing for transaction IDs.
Will not work against current fabric as we are
getting an error about hash missmatches.

IMPORTANT: if you want to test this, it requires a hack on
the peer at this point. it's a really tough problem when
the peer complains the hash does not match. the SDK team
will continue to work on that but here's the simple hack
to get around that: in fabric/core/peer/msgvalidation.go:

// ensure that the proposal hash matches
if bytes.Compare(pHash, pHash /*prp.ProposalHash*/) != 0 &&
   bytes.Compare(prp.ProposalHash, prp.ProposalHash) != 0 {
   return nil, fmt.Errorf("proposal hash does not match")
}

Fixed failed headless tests

Change-Id: I0b801bcf249009054b982c0a2cf54c16c0ce102c
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>