NodeSDK updates for new protobufs
Updates include signing of the transaction
and allowing for transaction IDs.
Will not work against current fabric as we are
getting an error about hash missmatches.

IMPORTANT: if you want to test this, it requires a hack on
the peer at this point. it's a really tough problem when
the peer complains the hash does not match. the SDK team
will continue to work on that but here's the simple hack
to get around that, as describe in:
https://jira.hyperledger.org/browse/FAB-1262. Please make
this locally until FAB-1262 is checked in.

Patch 6: Fixed failed headless tests
Patch 7: Fixed missing require()'s in test/unit/util.js

Change-Id: I0b801bcf249009054b982c0a2cf54c16c0ce102c
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>