FAB-1263 ECDSA signature malleability resistance
This change-set introduces ECDSA Signature malleability resistance.
ECDSA signatures do not have unique representation and this can facilitate
replay attacks and more. In order to have a unique representation,
this change-set forses BCCSP to generate and accept only signatures
with low-S.

Bitcoin has also addressed this issue with the following BIP:
https://github.com/bitcoin/bips/blob/master/bip-0062.mediawiki
Before merging this change-set, we need to ensure that client-sdks
generates signatures properly in order to avoid massive rejection
of transactions.

This is a port of the GO implementation here:
https://gerrit.hyperledger.org/r/#/c/2983

Change-Id: Iee78ee93f83ddfdd99526ea3cca9c11b33af8318
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>