Enhance importKey() to support private keys
[FAB-1811] support importing private keys from a PKCS#8 plain
text PEM file, and saves it in the keystore

This is needed to properly support bootstrapping an SDK from
pre-existing keys and certificates, because the SDK only knows how to 
enroll with a particular CA server, fabric-ca. For networks that are 
set up with other CAs, they crypto materials need to be generated by 
whatever tools or processes that's existing outside of SDK and 
fabric-ca, and the SDK must be able to load them in.

Change-Id: I87dadd33355623981387f28269d95d8ec1416303
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>