Fix broken test cases
FAB-1950
couchdb-fabriccop-tests.js treats an async API
User.setEnrollment() as synchronous API and as a result is not
getting the expected 'member' object.

cloudant-fabriccop-tests.js has the 'member' variable in the
callback function of the 'then' block following the
User.setEnrollment() call, and because it doesn't return the
member object, it results in the 'member' variable to be set the
'undefined' value.

Change-Id: I97a32b892803063acfef3074da54e4346cddf650
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>