Fix and rename cloudant and couchdb-fabriccop-tests
This change fixes a bug in User.setEnrollment where
a return is needed for this._signingIdentity to
allow a promise to be returned.  Also need to
change the cloudant and couchdb-fabriccop-tests to
expect an enrollment back from setEnrollment instead
of a user context.

Includes renaming of cop to ca in both the code as
well as the file names and .gitignore hfc/hfc-cop
to fabric-client/fabric-ca-client..

Change-Id: Ie387bde51003549416acd747c9ce962a9a7f17a3
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>