[FAB-2060] Transmit chaincodePath during deployment
We need to transmit the chaincode path to the peer now that
the peer builds our dockerfile.  The path is used for GOLANG
chaincode to understand the package that must be compiled.

In addition, packageChaincode() now produces a tar.gz that
contains the 'src/<package name>' directory structure instead
of the flattened files as before.

Fixes FAB-2060

Change-Id: Iff55c481fe0caf1061d61d51ea57fcf9b73a8222
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>