Refactor headless-tests.js into individual files
FAB-2164
this file is getting very big with thousands of lines of code.
it becomes increasingly difficult to code and debug with it.

this changeset breaks it apart into separate "test suites"
and group them together in a single folder dedicated to
headless tests (those that don't require a target network to run)

the test runner 'tape' does not support test suite hierarchy
(istanbul will not be able to find coverage information if a test
module is required from another module). so the solution is to use
a "gulp" task to launch it

Change-Id: I70a5afca13de10099989905d0f6df6db40cc8063
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>