Use mixin to enforce CryptoKeyStore APIs
FAB-2217 - step1
Goal is to allow CryptoKeyStore's API getKey() and putKey() to
be supported on top of any implementation of the KeyValueStore.
This is done by using a mixin layer to enforce the APIs on top
of the KeyValueStore implementation as the parameterized super-
class.

Usage of this new approach, in utils.getCryptoSuite() for instance
will be added in a later changeset.

Change-Id: I97cd8393ccd2882f505a33405c3a2797dfec4b1f
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>