Hash algorithms for signing and txId
This fixes the "invalid signature" problem when submitting
proposals to the peer. Turns out the peers are now using SHA3 256
which is a recent change.

In addition, the txID calculation always uses SHA2 256 so the code
around that must be changed to not be determined by the config
setting but instead always use SHA2 256.

Change-Id: I3538e1e388fb7d0a184e508c5792c659ecf91847
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>