nodeSDK Fix gulp test
'gulp test' gets an error because the endorser-tests
and install.js are picking up incorrect values set into
memory when cryptosuite-ecdsa-aes.js was running.  This
change sets the Config.js nconf, which is stored
as this._config, memory store to null, allowing the
default json file store to be used.

Change-Id: I77c80498c4494bc44cdf6125d027a78d580065db
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>