NodeSDK Add verify proposal response method - part 1
Add utiliy methods to Chain that will allow users to
verify the endorsement results. Verification will check
that that all write sets are the same between each
of the responses. Will check the signature to be sure
it was signed by the endorsing peer. And in part 2
it will validate the trust of the endorsing peer.

Change-Id: If4f626b6c38abba629b8f40bf4b821dbda4f8a2a
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>