nodeSDK Set logging level to debug, use default logger
When setting the debug logging level using the winston
default logger, setConfigSetting for 'hfc-logging' as
"debug":"console" needs to be done prior to getLogger
and prior to any other require that includes getLogger
and does not set the debug logging level.

In addition, changes are included to enable
couchdb-fabricca and cloudant-fabricca tests to run
from 'gulp test', and to store the kvs and crypto
options to the stateStore.  Some minor changes to fix
incorrect config settings were needed to enable all
the tests to run again.

Note that fabric-ca-services is not currently running
as part of 'gulp test' because of errors.  Will need
help fixing this test.

Change-Id: If8105c8bd2b30f9fff77578e7629370cca8cb851
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>