[WIP][FAB-2555] NodeSDK - Decode queried block
The queried block is a protobuf object with many fields
that are bytes, these need to be decoded into the objects
they represent. The new object will be returned that is
a pure JSON object with all fields converted from bytes
to values where possible.
Currently only the 'ENDORSER_TRANSACTION' are decoded.
I think we should decode the config'CONFIG' and 'CONFIG_UPDATE'
type blocks also.

Change-Id: Ife97e4b94db245039e317a5484f114c0e2b4d39a
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>