WIP[FAB-3090] NodeSDK - add APIs for create channel
This is a squashed commit of 3 separate fixes all needed to make
the build work again:

FAB-3090 NodeSDK - add APIs for create channel
The cryptotxgen tool is broken at the moment so the e2e test has
to rely on the SDK's new channel creation support now.

FAB-3277 events.js test broken and doesn't return
EventHub instances are not propertly cleaned up before the test
shuts down and as a result it will cause the test to hang.

FAB-3286 use an Admin identity to join channel
https://gerrit.hyperledger.org/r/#/c/7651/ enhanced the Join
Channel policy to require admin principals

Change-Id: I897a448d34cbac23245883f66b47271e7fee210c
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>