FAB-2923 Use cryptoSuite param instead of options
If an app needs special settings for the cryptoSuite,
the app can create the cryptoSuite on the User,
user newCryptoSuite(setting, KVSImplClass, opts),
then pass the cryptoSuite to FabricCAClient instead
of passing the options needed to create the cryptoSuite.

If no special settings are needed, the cryptoSuite
and state store will be created using default options.

Change-Id: I6f682d531fe70ce6d29790c0d35b06956e12441f
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>