[FAB-2787] NodeSDK - Client timeout
Add a heartbeat to the EventHub. The heartbeat
is using a new keep_alive event type added to the
Peer event service that will echo the keep_alive
request back as the response.  This will provide
traffic over the connection during periods when
there is no event traffic.
Note: this does require a proto change on the fabric side.
see https://gerrit.hyperledger.org/r/#/c/9971/

Change-Id: Ib4b24a1e2407790d31e2838a83d61bd2bd16afd9
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>