[WIP] [FAB-2843] NodeSDK - Handle network issues Part 3
Add a heartbeat to the EventHub. The heartbeat
is using a new keep_alive event type added to the
Peer event service. The will will just echo the
keep_alive call back as the response.  This will provide
traffic over the connection during periods when there is no
event traffic.
Note: this does require a proto change on the fabric side.
see [put gerrit number here]

Change-Id: Ib4b24a1e2407790d31e2838a83d61bd2bd16afd9
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>