[FAB-5048-medium] NodeSDK - add to package.json
Add two packages to the fabric-client Node.js
dependencies, 'long' and 'nano'.
This is a medium jira with a low risk of including.
There is an easy workaround for users if they see
this issue. This issue will not cause any runtime
problem, as the node service will not start if the
dependency is missing, which then can be easily
installed.

Change-Id: I09972d3fe8c6dd955089812436ec9f5c3649570d
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>