[WIP] [FAB-5363] NodeSDK - config part 1
This is a work in progress on the NodeSDK use of the
common network configuration. This is using the
latest yaml schema as described in the FAB. This
demonstrates how the SDK would load a configuration
and build a channel instance based on the network
configuration. This part only addresses channel,
orderer, and peer.

 API's that have been explored:
  getChannel()
  createChannel()
  joinChannel()

Change-Id: I0b1479d12684aa13bc052e872fbf118166004453
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>