[FAB-8373] shrinkwrap to control package versions
Adding npm-shrinkwrap.json in order to control dependent package
versions.  In this case, we install version 5.0.3 of the hoek
package in order to fix a security vulnerability flagged by nsp
(Node Security Platform).  Although this causes npm to install
version 5.0.3 of hoek at all places, 'nsp check' in the
fabric-sdk-node/fabric-client directory still shows hoek vulnerabilities.
This, however, appears to be a problem with the nsp tool and is a
false positive.

Change-Id: I6ea05ee4cf8970bd990a5fedfe102fefa0cf2dda
Signed-off-by: Keith Smith <bksmith@us.ibm.com>