[FAB-7670] introduce new API executeTransaction()
Application developers need to write many lines of code to make
their applications to execute transactions in Fabric network,
even if they use fabric SDKs because they only provides APIs for
each step in the transaction sequence (explained in docs/source/txflow.rst.)

So this patch is suggesting an API which executes whole transaction
lifecyle in a single call. With this interface, an application just
submits transaction and gets the result from Fabric network that
shows the transaction is commited or not.

Change-Id: I414f846fc49df454648ab64b21331479bcb08355
Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
---
ChangeLog (v1 -> v2):
- fix error in fabric-client/types/index.d.ts (thanks to Zhao Chaoyi)
- rename execTransaction to executeTransaction
- fix unit test failure due to invalid channel name with whitespaces