FAB-10917 NodeSDK - return channel methods
Peer roles are channel based and as such needed to be
set on Peer per channel. A new object has been introduced
to hold the peer and information about the peer, ChannelPeer.
This object will be unique per channel as a Peer may be on many
Channels. User will not get ChannelPeer objects on the
channel.getPeers() call. The ChannelPeer will wrapper the Peer
calls such that a ChannelPeer may be used as a Peer. The ChannelPeer
will also have the isInRole and setRole methods that were incorrectly
placed on the Peer. Having the ChannelPeer will make it easier to get
the ChannelEventHub, just use the getChannelEventHub() method on the
ChannelPeer object.

Change-Id: I7b718530805af99336d98210db99f1c46f2bc4e5
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>