FAB-10916 NodeSDK - allow create peers without TLS
The addition of Discovery has required a client cert/key pair
when remotes enpoints are created.  When not provided a self
cert key pair will be created based off the current user.
V1.0 allowed remote endpoints to be created before assigning
a user to the client and/or assigning a client cert/key pair.
This change will not require the assigning of cert/key or user
for remote endpoints to be created.

Change-Id: I237da5a030b1e9a0b379706239e9cce070331186
Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>