[FABN-830] Client calls toBytes of PKCS11 keys
When using an HSM, the _setUserFromConfig() method of
fabric-client/lib/Client.js calls the toBytes() method of the private key.
In the case of an HSM this operation isn't allowed and it throws an error.

This change checks to see if the key is a PKCS11 key and constructs the
cryptoContent accordingly.

Change-Id: Iaae0ce0fca0316c252a851393770636a4c7cf42a
Signed-off-by: Jonathan Patchell <Jonathan.Patchell@gemalto.com>