Fix issue installing chaincode package
Client.installChaincode is supposed to
support installing both from source (which
creates a chaincode package behind the scenes)
as well as installing an existing package.

The code was wrapping an existing package
as as a ChaincodeDeploymentSpec rather than
just using the package bytes directly when
invoking install.

This fixes that issue and adds a unit test
as well.

Fixes FABN-855

Change-Id: I3868327908dac4ff4e8e918247cb0f7d61de951a
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>