[FABN-976] Fix chaincode package install arg validation
Client.installChaincode requires a chaincode ID/version with chaincode
package. This is wrong, as the chaincode ID/version are not actually used
when you specify a chaincode package. The combo of required options
should be:

chaincodeId, chaincodeVersion, and chaincodePath
-or-
chaincodePackage

Also fixed the TypeScript definitions to allow for chaincode package
install without needing a chaincode ID/version, and did a bit of a
rework on the unit tests to cope with these changes - like ensuring
that errors are actually thrown!

Change-Id: I9632ed6b8897c2c3a6a71ce1db674b3f0e59e9a7
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>