FABN-1083 NodeSDK allow commit/endorse handler
Allow the commit handler and the endorsement handler
to be used when the channel is not initialized.
The sendTransaction method will be enhanced to create
and initialize the defined commit handler.
The sendTransactionProposal method will be enhanced to
create and initialize the defined endorsement handler.
Both methods will now call the handlers if defined.
Handlers may not be initialzed in the channel
constructor as it is not an async call.

Change-Id: I579ecfe446886f92380e5ec125554d7147d4da43
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>