[FABN-1130] Stop using "init" as default function name
The SDK currently forces you to supply a function name
to the instantiate command; if you don't, it defaults
to "init". There is no need to specify a function name
during instantiate, and the new programming model does
not need you to call *any* function at instantiate time.

This is the v1.4 backport; in order to maintain compat
with existing SDK applications, this change will keep
the current behaviour if you don't pass "fcn", but will
change the behaviour if you pass "fcn" with a null or
empty string value (highly unlikely anyone is doing that).

Change-Id: Ia526d855f66fc428bca43786bc4555e9da074a41
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>