[FABN-1159] Allow PEM paths relative to CCP
Allow a connection profile to specify paths to PEM files (for peer,
orderer, and certificate authority  TLS CA certs) that are relative
to the connection profile file. Currently they must be relative to
the current working directory, which is kind of useless if you want
to zip up a connection profile and the PEM files and ship them around.

Change-Id: Icd23c90553e5e4c81f6fe9245b6740b83749ec3e
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>