[FABN-1160] Fix TLS logic for discovered peers
The SDK assumes by default that peers and orderers
returned by service discovery will be running with
TLS (grpcs://). This is a bad assumption when we
are connecting to a gateway peer that is not running
with TLS (grpc://).

Change the logic so that if we connect to a peer
to perform discovery over TLS, then we assume that
the discovered peers and orderers also use TLS.
Likewise if we connect to a peer to perform discovery
without TLS, then assume that the discovered peers
and orderers also do not use TLS.

It is highly unlikely (impossible even?) that you
can have a mixed Fabric network with TLS and non-TLS
peers.

Change-Id: I9b8e0056cd6c6c743d6d7feab561cdaa8f7d3f05
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>