[WIP] FABN-1347 NodeSDK fabric-base initial
Initial files added to new module fabric-base.
Some unit test and an integration test
showing the complete flow of an endorsement, commit with
monitoring for an event, query and discovery calls.

Note that the channel class is not the focal point
of the flow, the new class "Proposal" is the object that
is used to build up the proposal, sign it, and then
send it. Then again to build up the commit submission,
using the results of the endorsement, sign it, then send
it to the orderer. The "Client" class is still the central
location for client side shared connection settings like the
client side mutual TLS. A new class "TransactionContext" is
applied at each building of a proposal, avoiding the client
class having user state and a state or crypto store.
New ChannelDiscovery class added to be the focal point of discovery
calls to the peer for a channel. This class will build
and sign the discovery request. This class will
decode the discover configurations and
endorsement plans. It will provide peers and orderers to
the channel along with the plan for the endorsement hsndler,
query handler, and commit handler. The dicovered peers and
organizations will be the source endpoint for event monitoring.

Change-Id: If7dc29aaab2697d4979b36f209a23bfc43588d56
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>