[WIP] FABN-12xx NodeSDK-Base initial endorse flow
For viewing only, initial files added to new module
fabric-base. Some unit test and an integration test
showing the complete flow of an endorsement.

Note that the channel class is not the focal point
of the flow, the new class "Proposal" is the object that
is used to build up the proposal, sign it, and then
send it. Then again to build up the commit submission,
using the results of the endorsement, sign it, then send
it to the orderer. The "Client" class is still the central
location for client side shared connection settings like the
client side mutual TLS. A new class "TransactionContext" is
applied at each building of a proposal, avoiding the client
class having user state.

No discovery, common connectionn profile, or events yet.

Change-Id: If7dc29aaab2697d4979b36f209a23bfc43588d56
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>