[FABN-1341] Support .fabricignore files in packaging
The packaging code in the SDK for Fabric v1.4 lacks
general purpose options for controlling which files
in a directory get packaged into a CDS file. Node.js
has the option to handle a .npmignore file, but the
other platforms do not have the same option. This is
a particular problem with Java, because the build
directories containing uber JARs end up in the CDS
package, then get dropped the other end.

This CR adds support for an optional .fabricignore
file that can be placed in the contract directory.
This file can be used to ignore any files desired
by the user. In the case of Go and Node, which
ignore non-Go files and node_modules respectively,
this behaviour is disabled in the presence of a
.fabricignore file, giving the user full control
over what gets packaged.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I362d786733e2d8f8b12406c20ef02b66b1c6787d