FABN-1345 NodeSDK Add DeliverWithPrivateData support
the connect function in the ChannelEventHub have a new parameter in its
options called private_data, when set to true, it will deliver blocks
with private data. The new return willl contain a new atribute called
private_data, that will contain the set of writes of the private data
collections a peer can access.

Signed-off-by: Alexis Ortiz <alexis.ortiz1@ibm.com>
Change-Id: I5ddc155da0abbf0c21b77baace3d2e40054065c8