[FABN-1379] Fix dodgy E2E chaincode example_cc1
10:34:31  not ok 238 Error: Failed to test for bad
transient map. The chaincode should have rejected
the upgrade proposal.

This error is coming from a test that checks to see
if we get an upgrade failure when we send in bad transient
data.

Putting the stupid test aside, this E2E test uses a
Node.js chaincode that accesses an undocumented "map"
property of the transient map, which is no longer
present since the Node.js chaincode got upgraded to
protobufjs v6.

Fix the chaincode, but come back and kill the tests
soon.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I19e1b4b045fb75da339a9cf092604590d7542f06