FABN-1387: Fix fabric-network TypeScript definitions
index.d.ts referred to TypeScript source files not included in the
published npm module.

In the build, run a TypeScript compilation pass to generate definition
files in the output directory that gets published to npm. Ensure the
index.d.ts only references generated TypeScript definitions in the
output directory and not source files directly.

Change-Id: I11ab8154182a026d3c7a459108f4277c54dc4f15
Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>